added utils and new players test, authorisation test fixed
